requirements.txt
sdmx1
git+git://github.com/dougmet/yamlmd
git+git://github.com/brockfanning/sdg-build-1@un-sdmx-sdgs-api
